Breakup HostRuntimeEnvironment
Move the different types into their own source files.
