+++ src/Test/Utilities/Desktop/TestUtilities.Desktop.csproj
@@ -100,6 +100,10 @@
     <Compile Include="CLRHelpers.cs" />
     <Compile Include="ConditionalFactAttribute.cs" />
     <Compile Include="ProcessUtilities.cs" />
+    <Compile Include="Remote\Exceptions.cs" />
+    <Compile Include="ModuleData.cs" />
+    <Compile Include="Extensions.cs" />
+    <Compile Include="Remote\RuntimeAssemblyManager.cs" />
     <Compile Include="SharedResourceHelpers.cs" />
     <Compile Include="CommonDiagnosticAnalyzers.cs" />
     <Compile Include="CommonTestBase.CompilationVerifier.cs" />
